refactor: 'bottari' 패키지명을 'personal'로 변경
