+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/home/personal/BottariUiEvent.kt
@@ -1,4 +1,4 @@
-package com.bottari.presentation.view.home.bottari
+package com.bottari.presentation.view.home.personal
 
 sealed interface BottariUiEvent {
     data object BottariDeleteSuccess : BottariUiEvent
