+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/home/personal/BottariUiState.kt
@@ -1,4 +1,4 @@
-package com.bottari.presentation.view.home.bottari
+package com.bottari.presentation.view.home.personal
 
 import com.bottari.presentation.model.BottariUiModel
