+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/home/personal/BottariViewModel.kt
@@ -1,4 +1,4 @@
-package com.bottari.presentation.view.home.bottari
+package com.bottari.presentation.view.home.personal
 
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.initializer
