+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/home/personal/adapter/BottariViewHolder.kt
@@ -1,4 +1,4 @@
-package com.bottari.presentation.view.home.bottari.adapter
+package com.bottari.presentation.view.home.personal.adapter
 
 import android.graphics.Color
 import android.graphics.drawable.ClipDrawable
