+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/home/personal/create/BottariCreateUiEvent.kt
@@ -1,4 +1,4 @@
-package com.bottari.presentation.view.home.bottari.create
+package com.bottari.presentation.view.home.personal.create
 
 sealed interface BottariCreateUiEvent {
     data class CreateBottariSuccess(
