+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/home/personal/create/BottariCreateUiState.kt
@@ -1,4 +1,4 @@
-package com.bottari.presentation.view.home.bottari.create
+package com.bottari.presentation.view.home.personal.create
 
 data class BottariCreateUiState(
     val bottariTitle: String = "",
