+++ android/Bottari/presentation/src/main/res/layout/fragment_profile.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:focusable="true"
     android:focusableInTouchMode="true"
-    tools:context=".view.home.bottari.BottariFragment">
+    tools:context=".view.home.personal.BottariFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
